requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
openai>=0.27.0
lancedb>=0.1.0
fastembed>=0.1.0
pydantic>=2.0.0
httpx>=0.23.0
python-multipart>=0.0.5
matplotlib>=3.9.0
opentelemetry-api>=1.12.0
opentelemetry-sdk>=1.12.0
opentelemetry-instrumentation-fastapi>=0.48b0
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1
pytest-cov>=4.1.0
slowapi>=0.1.8
